requirements.txt
Django==1.9.8
django-ipware==1.1.5
dj-static==0.0.6
